authentication jwt and middlewares
